#4 Populate Version Dropdowns with CDN versions
When selecting a CDN you now get all the versions associated with the CDN to choose from locally. This allows users to specify versions when saving their choices in the admin menu. When switching CDN providers -- users will have to wait until the page reloads to get the new dropdown list associated with the CDN.